+++ lib/redis-throttle.rb
@@ -1,4 +1,3 @@
 # frozen_string_literal: true
 
 require_relative "./redis_throttle"
-require_relative "./redis/throttle"